Add support for Talos OS
Resolves https://github.com/metal-stack/csi-driver-lvm/issues/86

`/etc` on Talos OS is read-only, which means that this chart will fail
to install as is. The new `hostWritePath` option allows the default
`/etc/lvm` to be overwritten. For Talos OS, setting `hostWritePath:
/var/etc/lvm` fixes the install issue. This is just half the story.

After this installs, creating a PVs will fail since the controller has
some hard-coded paths starting with `/etc`. Here is the second half of
this fix: https://github.com/metal-stack/csi-driver-lvm/pull/87

First step is to get that PR 👆 merged first, then update the image
versions in this one before merging this change.

Learn more about the Talos' file system:
https://www.talos.dev/v1.4/learn-more/architecture/

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>