Added XID exhaustion monitoring
For a comprehensive description of the measurements refer to:

https://www.crunchydata.com/blog/managing-transaction-id-wraparound-in-postgresql

In short: we look after the oldest XID for reference. If it reaches 2
billion, XIDs will overflow and a SELECT on an affected table will
miss at least 50% of the rows. Also Postgres shuts down the database.

The percent_towards_wraparound value tells us how far away from that
point we are. If it reaches 100%, then we are dead. So, at ca 95% a
manual vacuum MUST be executed - AT LEAST on the affected table!

Last but not least, the percent_towards_emergency_autovac value tells
us, how far away from the next emergency autovacuum we are. This
happens, when autovacuum_freeze_max_age (by default 200 mio) is being
reached on a table. This is just informational as this does NOT lead
to immediate hazards when >100%.